requirements.txt
biopython>=1.83
streamlit>=1.35.0
